test: split out http2 from test-stream-pipeline
Splitting out the http2 portion of the test has a few benfits:

* We don't skip the rest of the tests if `node` is compiled without
  crypto.
* We can find out if the http2 portion of the test is responsible for
  the timeouts reported in issue 24456.

Refs: https://github.com/nodejs/node/issues/24456

PR-URL: https://github.com/nodejs/node/pull/24631
